implement non-returning message

+++ SocialMedia.Application/UserProfile/CommandHandlers/DeleteUserCommandHandler.cs
@@ -1,10 +1,12 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using SocialMedia.Application.Enums;
+using SocialMedia.Application.Models;
 using SocialMedia.Application.UserProfile.Commands;
 using SocialMedia.Dal.Data;
 
 namespace SocialMedia.Application.UserProfile.CommandHandlers;
-internal class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand>
+internal class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, OperationResult<Domain.Aggregates.UserProfileAggregate.UserProfile>>
 {
     private readonly DataContext _context;
 
@@ -13,18 +15,28 @@ internal class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand>
         _context = context;
     }
 
-    public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+    public async Task<OperationResult<Domain.Aggregates.UserProfileAggregate.UserProfile>> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
+        var result = new OperationResult<Domain.Aggregates.UserProfileAggregate.UserProfile>();
+
         var userProfile = await _context.UserProfiles.FirstOrDefaultAsync(user => user.UserProfileId == request.UserProfileId, cancellationToken);
-        
+
         // TODO: Return type will be decided later
-        //if(userProfile is null)
-        //{
-        //    return new Unit();
-        //}
-        
+        if (userProfile is null)
+        {
+            result.IsError = true;
+            Error error = new()
+            {
+                Code = ErrorCode.NotFound,
+                Message = "No User is Found"
+            };
+            result.Errors.Add(error);
+            return result;
+        }
+
         _context.UserProfiles.Remove(userProfile);
         await _context.SaveChangesAsync(cancellationToken);
-        return new Unit();
+        result.Payload  = userProfile;
+        return result;
     }
 }
